+++ src/input.c
@@ -5,7 +5,7 @@
  * 1: Line complete
  */
 
-unsigned int kinput(unsigned int maxlen, console* cons)
+size_t kinput(size_t maxlen, console* cons)
 {
 	cons->last = kgetc(cons);
 	
@@ -18,7 +18,6 @@ unsigned int kinput(unsigned int maxlen, console* cons)
 		cons->internalbuffer = (char*)kmalloc(maxlen + 1);
 		cons->buffer = cons->internalbuffer;
 		cons->bufcnt = 0;
-		//kprintf("new buf");
 	}
 
 	switch (cons->last) {
